requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect==1.2.0
cssutils==2.11.1
ecdsa==0.19.0
emails==0.6
fastapi==0.115.6
fastapi-pagination==0.12.32
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.5.0
passlib==1.7.4
premailer==3.10.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
pytz==2024.2
requests==2.32.3
rsa==4.9
sentry-sdk==2.19.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
vine==5.1.0
wcwidth==0.2.13
